Randomized the coloring of the board cells
This is to help them better identify what pattern is being worked on.
Children tend to learn words by memorizing the font size and shape and
thus make us believe they know the word when they really do not. And
by reenforcing the idea to them that they do know the word it creates a
frustration when they are unable to identify it later due to it not
being the same pattern.

By randomizing everything, they are able to "refine" the pattern in
the mind to help eliminate certain aspects of the pattern as being the
word itself allowing for a better and more accurate learning process.

Still left to do is the randomization of the font style and size of the
fonts to help eliminate those aspects of the word learning association.