Ask questions in a variety of styles
The font style, color, and size changes up to help prevent visual
memorization of the asked questions.

Added a player level variance to reduce/increase the varienty of unique
questions to be asked. This helps to repeat the same questions multiple
times to help reinforce the understanding of the answers during a play
for the younger kids.